Move converters and timer to utils/ from core
