+++ bot/cogs/moderation/slowmode.py
@@ -2,7 +2,7 @@ from discord.ext.commands import BadArgument, Cog, Context, command
 from loguru import logger
 
 from bot.core.bot import Bot
-from bot.core.converters import Duration
+from bot.utils.converters import Duration
 from bot.utils.time import stringify_duration
